:truck: Move feature modules to new feature_mdoules folder
